backend: move Page struct to dto module
